backend-ai-services/mediapipe/requirements.txt
mediapipe>=0.10.7
opencv-python-headless>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.25.0
loguru>=0.7.0
prometheus-client>=0.18.0
psutil>=5.9.0
tensorflow>=2.15.0
